backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
mysqlclient==2.2.4
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
Pillow==10.3.0
itsdangerous==2.2.0
requests==2.32.3

